LayoutTests/media/track/track-load-from-src-readyState-expected.txt

Tests the load event on HTMLTrackElement and LOADED readyState on TextTrack when src is set from JavaScript.

EXPECTED (track.track.readyState == '0') OK
EVENT(load)
EXPECTED (track.track.readyState == '2') OK
END OF TEST
